StreetMetadata: split geotag into its own component
