+++ assets/scripts/streets/StreetMetaData.jsx
@@ -3,13 +3,12 @@ import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 import { IntlProvider, FormattedMessage } from 'react-intl'
 import { formatDate } from '../util/date_format'
-import { trackEvent } from '../app/event_tracking'
 import { t } from '../app/locale'
 import { getRemixOnFirstEdit } from './remix'
 import { showGallery } from '../gallery/view'
 import StreetWidth from './StreetWidth'
+import StreetMetaGeotag from './StreetMetaGeotag'
 import Avatar from '../users/Avatar'
-import { SHOW_DIALOG } from '../store/actions'
 
 class StreetMetaData extends React.Component {
   static propTypes = {
@@ -17,8 +16,6 @@ class StreetMetaData extends React.Component {
     signedIn: PropTypes.bool.isRequired,
     userId: PropTypes.string,
     street: PropTypes.any,
-    enableLocation: PropTypes.bool,
-    showGeotagDialog: PropTypes.func,
     locale: PropTypes.object
   }
 
@@ -32,16 +29,6 @@ class StreetMetaData extends React.Component {
     this.state = {
       street: this.props.street
     }
-  }
-
-  onClickGeotag = (event) => {
-    event.preventDefault()
-    if (!this.props.street.location) {
-      trackEvent('Interaction', 'Clicked add location', null, null, true)
-    } else {
-      trackEvent('Interaction', 'Clicked existing location', null, null, true)
-    }
-    this.props.showGeotagDialog()
   }
 
   onClickAuthor = (event) => {
@@ -68,33 +55,6 @@ class StreetMetaData extends React.Component {
     )
   }
 
-  getGeotagText = () => {
-    const { hierarchy } = this.props.street.location
-    const unknownLabel = t('dialogs.geotag.unknown-location', 'Unknown location')
-    let text = ''
-    text = (hierarchy.locality) ? hierarchy.locality
-      : (hierarchy.region) ? hierarchy.region
-        : (hierarchy.neighbourhood) ? hierarchy.neighbourhood
-          : unknownLabel
-    if (text !== unknownLabel && hierarchy.country) {
-      text = text + ', ' + hierarchy.country
-    }
-    return text
-  }
-
-  renderGeotag = (street, readOnly) => {
-    const geotagText = (street.location) ? this.getGeotagText() : t('dialogs.geotag.add-location', 'Add location')
-    const geolocation = (
-      <span className="street-metadata-map">
-        { (readOnly) ? geotagText : (
-          <a onClick={this.onClickGeotag}> {geotagText} </a>
-        ) }
-      </span>
-    )
-
-    return (readOnly && !street.location) ? null : geolocation
-  }
-
   render () {
     let author = null
     const creatorId = this.props.street.creatorId
@@ -103,8 +63,6 @@ class StreetMetaData extends React.Component {
     } else if (!creatorId && (this.props.signedIn || getRemixOnFirstEdit())) {
       author = t('users.byline', 'by {user}', { user: t('users.anonymous', 'Anonymous') })
     }
-
-    const geolocation = (this.props.enableLocation) ? this.renderGeotag(this.props.street, this.props.readOnly) : null
 
     return (
       <IntlProvider
@@ -116,7 +74,7 @@ class StreetMetaData extends React.Component {
           <StreetWidth readOnly={this.props.readOnly} />
           <span className="street-metadata-author">{author}</span>
           <span className="street-metadata-date">{formatDate(this.props.street.updatedAt)}</span>
-          {geolocation}
+          <StreetMetaGeotag />
         </div>
       </IntlProvider>
     )
@@ -127,20 +85,8 @@ function mapStateToProps (state) {
   return {
     signedIn: state.user.signedIn,
     userId: state.user.signInData && state.user.signInData.userId,
-    enableLocation: state.flags.GEOLOCATION.value,
     locale: state.locale
   }
 }
 
-function mapDispatchToProps (dispatch) {
-  return {
-    showGeotagDialog: () => {
-      dispatch({
-        type: SHOW_DIALOG,
-        name: 'GEOTAG'
-      })
-    }
-  }
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(StreetMetaData)
+export default connect(mapStateToProps)(StreetMetaData)
